Feat
Create Customer service, getCustomer function and create function on service.
